// frontend/src/types/ServerListRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ServerListRequest = { limit: number, offset_id: number | null, licensed: boolean | null, checked: boolean | null, spoofable: boolean | null, crashed: boolean | null, has_players: boolean | null, online: boolean | null, };